# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "学员端(hrwai_user) 学习/练习/考试/收藏/搜索/资料/论坛等 P0/P1 能力（ADR-0017/0018）。鉴权：`Authorization: Bearer \u003caccess JWT\u003e`（access 2h，`POST /api/auth/refresh` 换新双令牌）。响应统一 `{code,message,data}`（见 ADR-0005）。本页由 gin-swagger 生成，范围 B（~30 端点）。",
        "title": "叉车维修培训系统-学员端 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/level-exam/available": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "学员可参加的定级考试场次",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-定级考试"
                ],
                "summary": "定级考试-可用列表",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/forklift-training_internal_service.LevelExamAvailableDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
        "/level-exam/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页查询学员定级考试历史",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-定级考试"
                ],
                "summary": "定级考试历史",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.LevelExamHistoryDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
        "/level-exam/participants/{participant_id}/result": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "查询参与记录的判分结果",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-定级考试"
                ],
                "summary": "查询结果",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "参与记录ID",
                        "name": "participant_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.LevelExamResultDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    },
                    "404": {
                        "description": "不存在",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
        "/level-exam/participants/{participant_id}/save": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "保存答题进度与剩余时间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-定级考试"
                ],
                "summary": "保存答案",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "参与记录ID",
                        "name": "participant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "答案",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
        "/level-exam/participants/{participant_id}/submit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "提交试卷（支持 is_timeout/remaining_time）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-定级考试"
                ],
                "summary": "交卷",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "参与记录ID",
                        "name": "participant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "交卷",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.LevelExamParticipantDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
        "/level-exam/sessions/{session_id}/enter": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "进入指定场次并获取试卷与参与记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-定级考试"
                ],
                "summary": "进入定级考试",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "场次ID",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.LevelExamDataDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "forklift-training_internal_service.LevelExamAvailableDTO": {
            "type": "object",
            "properties": {
                "can_enter": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "duration": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "has_participated": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "participant_id": {},
                "participant_status": {},
                "participants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forklift-training_internal_service.LevelExamParticipantDTO"
                    }
                },
                "pass_score": {
                    "type": "integer"
                },
                "question_config": {},
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total_score": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "forklift-training_internal_service.LevelExamHistoryDTO": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forklift-training_internal_service.LevelExamParticipantDTO"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "forklift-training_internal_service.LevelExamParticipantDTO": {
            "type": "object",
            "properties": {
                "answers_snapshot": {},
                "created_at": {
                    "type": "string"
                },
                "exam_session_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_passed": {
                    "type": "boolean"
                },
                "objective_score": {
                    "type": "number"
                },
                "question_ids": {},
                "remaining_time": {
                    "type": "integer"
                },
                "score": {
                    "type": "number"
                },
                "session_name": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "student_id": {
                    "type": "integer"
                },
                "student_name": {
                    "description": "列表路径附带：学员姓名 / 场次名称（未加载时省略）。",
                    "type": "string"
                },
                "subjective_score": {
                    "type": "number"
                },
                "submit_time": {
                    "type": "string"
                }
            }
        },
        "forklift-training_internal_service.LevelExamResultDTO": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forklift-training_internal_service.LevelExamAnswerDTO"
                    }
                },
                "participant": {
                    "$ref": "#/definitions/forklift-training_internal_service.LevelExamParticipantDTO"
                }
            }
        },
...
        "forklift-training_pkg_response.R": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer \u003caccess JWT\u003e，示例：Bearer eyJhbGciOi...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}